GROK-12702: JS API: Moved TagsInput to JS API
